Shrink probability remap tables.
Saves 2288 bytes in vp8+vp9 libvpx.a.

+++ vp9/decoder/vp9_dsubexp.c
@@ -29,7 +29,7 @@ static int decode_uniform(vpx_reader *r) {
 }
 
 static int inv_remap_prob(int v, int m) {
-  static int inv_map_table[MAX_PROB] = {
+  static uint8_t inv_map_table[MAX_PROB] = {
       7,  20,  33,  46,  59,  72,  85,  98, 111, 124, 137, 150, 163, 176, 189,
     202, 215, 228, 241, 254,   1,   2,   3,   4,   5,   6,   8,   9,  10,  11,
      12,  13,  14,  15,  16,  17,  18,  19,  21,  22,  23,  24,  25,  26,  27,
